Issue #145 - merge FeedFetcher and FeedUpdater, Issue #121 - Cleanup of its unit tests
